+++ src/glados/__init__.py
@@ -1,6 +1,6 @@
 """GLaDOS - Voice Assistant using ONNX models for speech synthesis and recognition."""
 
-from .engine import Glados, GladosConfig
+from .core.engine import Glados, GladosConfig
 
 __version__ = "0.1.0"
 __all__ = ["Glados", "GladosConfig"]
