--- src/glados/core/audio_message.py
@@ -1,11 +1,0 @@
-from dataclasses import dataclass
-
-import numpy as np
-from numpy.typing import NDArray
-
-
-@dataclass
-class AudioMessage:
-    audio: NDArray[np.float32]
-    text: str
-    is_eos: bool = False
